+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - feat: add support for slide deck localization
 - feat: add controls for localization
+- docs: add table of contents to README.md
 
 # 0.11.1